moved scss files into respective component folders
